+++ decidim-conferences/app/types/decidim/conferences/conference_type.rb
@@ -13,7 +13,7 @@ module Decidim
       description "A conference"
 
       field :subtitle, Decidim::Core::TranslatedFieldType, "The subtitle of this conference"
-      field :shortDescription, Decidim::Core::TranslatedFieldType, "The sort description of this conference", property: :short_description
+      field :shortDescription, Decidim::Core::TranslatedFieldType, "The short description of this conference", property: :short_description
       field :description, Decidim::Core::TranslatedFieldType, "The description of this conference"
       field :slug, types.String, "The slug of this conference"
       field :hashtag, types.String, "The hashtag for this conference"
@@ -21,7 +21,7 @@ module Decidim
       field :location, types.String, "The location of this conference"
       field :createdAt, Decidim::Core::DateTimeType, "The time this conference was created", property: :created_at
       field :updatedAt, Decidim::Core::DateTimeType, "The time this conference was updated", property: :updated_at
-      field :publishedAt, Decidim::Core::DateTimeType, "The time this conference was updated", property: :published_at
+      field :publishedAt, Decidim::Core::DateTimeType, "The time this conference was published", property: :published_at
       field :reference, types.String, "Reference prefix for this conference"
 
       field :heroImage, types.String, "The hero image for this conference", property: :hero_image
@@ -29,23 +29,16 @@ module Decidim
       field :promoted, types.Boolean, "If this conference is promoted (therefore in the homepage)"
       field :objectives, Decidim::Core::TranslatedFieldType, "The objectives of the conference"
       field :showStatistics, types.Boolean, "If this conference shows the statistics", property: :show_statistics
-      field :startDate, Decidim::Core::DateType, "The time this conference was created", property: :start_date
-      field :endDate, Decidim::Core::DateType, "The time this conference was updated", property: :end_date
+      field :startDate, Decidim::Core::DateType, "The date this conference starts", property: :start_date
+      field :endDate, Decidim::Core::DateType, "The date this conference ends", property: :end_date
       field :registrationsEnabled, types.Boolean, "If the registrations are enabled in this conference", property: :registrations_enabled
       field :availableSlots, types.Int, "The number of available slots in this conference", property: :available_slots
       field :registrationTerms, Decidim::Core::TranslatedFieldType, "The registration terms of this conference", property: :registration_terms
 
-      field :speakers, types[Decidim::Conferences::ConferenceSpeakerType], ""
-      field :partners, types[Decidim::Conferences::ConferencePartnerType], ""
-      field :categories, types[Decidim::Core::CategoryType], ""
-      field :conferenceRegistrations, types[Decidim::Conferences::ConferenceRegistrationType], "", property: :conference_registrations
-      field :mediaLinks, types[Decidim::Conferences::ConferenceMediaLinkType], "", property: :media_links
-      field :registrationTypes, types[Decidim::Conferences::ConferenceRegistrationTypeType], "", property: :registration_types
-
-      # These fields may be private:
-      # decidim_participatory_process_group_id
-      # decidim_organization_id
-      # private_space
+      field :speakers, types[Decidim::Conferences::ConferenceSpeakerType], "List of speakers in this conference"
+      field :partners, types[Decidim::Conferences::ConferencePartnerType], "List of partners in this conference"
+      field :categories, types[Decidim::Core::CategoryType], "List of categories in this conference"
+      field :mediaLinks, types[Decidim::Conferences::ConferenceMediaLinkType], "List of media links in this conference", property: :media_links
     end
   end
 end