+++ decidim-consultations/app/types/decidim/consultations/consultation_type.rb
@@ -13,10 +13,10 @@ module Decidim
 
       field :subtitle, Decidim::Core::TranslatedFieldType, "The subtitle of this consultation"
       field :description, Decidim::Core::TranslatedFieldType, "The description of this consultation"
-      field :slug, !types.String
+      field :slug, !types.String, "Slug of this consultation"
       field :createdAt, !Decidim::Core::DateTimeType, "The time this consultation was created", property: :created_at
       field :updatedAt, !Decidim::Core::DateTimeType, "The time this consultation was updated", property: :updated_at
-      field :publishedAt, !Decidim::Core::DateTimeType, "The time this consultation was updated", property: :published_at
+      field :publishedAt, !Decidim::Core::DateTimeType, "The time this consultation was published", property: :published_at
 
       field :introductoryVideoUrl, types.String, "The introductory video url for this consultation", property: :introductory_video_url
       field :introductoryImage, types.String, "The introductory image for this consultation", property: :introductory_image
@@ -27,10 +27,6 @@ module Decidim
       field :resultsPublishedAt, Decidim::Core::DateType, "Date when the results have been published", property: :results_published_at
 
       field :questions, types[Decidim::Consultations::ConsultationQuestionType], ""
-      # These fields may be private:
-      # decidim_participatory_process_group_id
-      # decidim_organization_id
-      # private_space
     end
   end
 end